+++ src/gui.h
@@ -50,13 +50,13 @@ void		init_gui(void);
 gint            delete_event(GtkWidget *, GdkEvent *, gpointer);
 void            toggle_effect(GtkWidget *, effect_t *);
 void            bank_append_entry(gchar *entry);
-void            set_vumeter_in_value(double power);
-void            set_vumeter_out_value(double power);
+void            set_vumeter_in_value(float power);
+void            set_vumeter_out_value(float power);
 void            gnuitar_printf(char *fmt, ...);
 
 extern GtkWidget *processor;
-extern double master_volume;
-extern double input_volume;
+extern float master_volume;
+extern float input_volume;
 extern GtkObject* adj_input;
 extern GtkObject* adj_master;
